refactor: formatNumberWithUnits 리팩토링
