+++ packages/utils/src/formatter/formatNumberWithUnits/formatNumberWithUnits.types.ts
@@ -1,0 +1,27 @@
+type FloorUnit =
+  | 1
+  | 10
+  | 100
+  | 1_000
+  | 10_000
+  | 100_000
+  | 1_000_000
+  | 10_000_000
+  | 100_000_000
+  | 1_000_000_000
+  | 10_000_000_000
+  | 100_000_000_000
+  | 1_000_000_000_000;
+
+export interface Unit {
+  unit: string;
+  value: number;
+}
+
+export interface FormatNumberWithUnitsOptions {
+  units?: Unit[] | readonly Unit[];
+  commas?: boolean;
+  floorUnit?: FloorUnit;
+  space?: boolean;
+  decimal?: number;
+}
